Get rid of different factory files
..and fix some cypress tests
