+++ backend/src/schema/resolvers/donations.spec.js
@@ -1,11 +1,10 @@
 import { createTestClient } from 'apollo-server-testing'
-import Factory from '../../factories'
+import Factory, { cleanDatabase } from '../../factories'
 import { gql } from '../../helpers/jest'
 import { getNeode, getDriver } from '../../db/neo4j'
 import createServer from '../../server'
 
 let mutate, query, authenticatedUser, variables
-const factory = Factory()
 const instance = getNeode()
 const driver = getDriver()
 
@@ -33,7 +32,7 @@ const donationsQuery = gql`
 describe('donations', () => {
   let currentUser, newlyCreatedDonations
   beforeAll(async () => {
-    await factory.cleanDatabase()
+    await cleanDatabase()
     authenticatedUser = undefined
     const { server } = createServer({
       context: () => {
@@ -50,11 +49,11 @@ describe('donations', () => {
 
   beforeEach(async () => {
     variables = {}
-    newlyCreatedDonations = await factory.create('Donations')
+    newlyCreatedDonations = await Factory.build('donations')
   })
 
   afterEach(async () => {
-    await factory.cleanDatabase()
+    await cleanDatabase()
   })
 
   describe('query for donations', () => {
@@ -68,7 +67,7 @@ describe('donations', () => {
 
       describe('authenticated', () => {
         beforeEach(async () => {
-          currentUser = await factory.create('User', {
+          currentUser = await Factory.build('user', {
             id: 'normal-user',
             role: 'user',
           })
@@ -102,7 +101,7 @@ describe('donations', () => {
       describe('authenticated', () => {
         describe('as a normal user', () => {
           beforeEach(async () => {
-            currentUser = await factory.create('User', {
+            currentUser = await Factory.build('user', {
               id: 'normal-user',
               role: 'user',
             })
@@ -121,7 +120,7 @@ describe('donations', () => {
 
         describe('as a moderator', () => {
           beforeEach(async () => {
-            currentUser = await factory.create('User', {
+            currentUser = await Factory.build('user', {
               id: 'moderator',
               role: 'moderator',
             })
@@ -140,7 +139,7 @@ describe('donations', () => {
 
         describe('as an admin', () => {
           beforeEach(async () => {
-            currentUser = await factory.create('User', {
+            currentUser = await Factory.build('user', {
               id: 'admin',
               role: 'admin',
             })
