+++ backend/src/schema/resolvers/passwordReset.spec.js
@@ -1,4 +1,4 @@
-import Factory from '../../factories'
+import Factory, { cleanDatabase } from '../../factories'
 import { gql } from '../../helpers/jest'
 import { getNeode, getDriver } from '../../db/neo4j'
 import createPasswordReset from './helpers/createPasswordReset'
@@ -7,7 +7,6 @@ import { createTestClient } from 'apollo-server-testing'
 
 const neode = getNeode()
 const driver = getDriver()
-const factory = Factory()
 
 let mutate
 let authenticatedUser
@@ -39,14 +38,14 @@ beforeAll(() => {
 })
 
 afterEach(async () => {
-  await factory.cleanDatabase()
+  await cleanDatabase()
 })
 
 describe('passwordReset', () => {
   describe('given a user', () => {
     beforeEach(async () => {
-      await factory.create(
-        'User',
+      await Factory.build(
+        'user',
         {},
         {
           email: 'user@example.org',
@@ -127,8 +126,8 @@ describe('resetPassword', () => {
 
   describe('given a user', () => {
     beforeEach(async () => {
-      await factory.create(
-        'User',
+      await Factory.build(
+        'user',
         {
           role: 'user',
         },
